fix: move tests to be able to use common
