+++ crates/server/src/tests/mod.rs
@@ -1,3 +1,4 @@
 mod auth;
+mod common;
 mod protected;
 mod users;
