+++ docs/todo.md
@@ -8,6 +8,7 @@
 - Websockets
 - Tracing/logging during unit tests
 - Move request structures to server
+- Remove sensitive data from tracing
 
 ## Devops
